refactor: move handler files to proxy/handler/ subdirectory
- openai, anthropic, proxy-handler -> proxy/handler/
- Fixed all import paths in proxy/ subdirectories, src/index.ts, tests
- proxy-core.ts, types.ts, proxy-logging.ts, log-helpers.ts remain at proxy/ top level (shared utilities)
- All 580 tests pass
