+++ src/index.ts
@@ -17,8 +17,8 @@ import { getConfig, Config } from "./config/index.js";
 import { initDatabase, getAllProviders, backfillMetricsFromRequestMetrics } from "./db/index.js";
 import { loadRecommendedConfig } from "./config/recommended.js";
 import { authMiddleware } from "./middleware/auth.js";
-import { openaiProxy } from "./proxy/openai.js";
-import { anthropicProxy } from "./proxy/anthropic.js";
+import { openaiProxy } from "./proxy/handler/openai.js";
+import { anthropicProxy } from "./proxy/handler/anthropic.js";
 import { adminRoutes } from "./admin/routes.js";
 import { RetryRuleMatcher } from "./proxy/orchestration/retry-rules.js";
 import { ProviderSemaphoreManager } from "./proxy/orchestration/semaphore.js";
