+++ src/proxy/handler/openai.ts
@@ -2,20 +2,20 @@ import type { FastifyPluginCallback, FastifyReply } from "fastify";
 import { randomUUID } from "crypto";
 import Database from "better-sqlite3";
 import fp from "fastify-plugin";
-import { getActiveProviders, insertRequestLog } from "../db/index.js";
-import { getSetting } from "../db/settings.js";
-import { decrypt } from "../utils/crypto.js";
-import { proxyGetRequest, createErrorFormatter, type ProxyErrorResponse } from "./proxy-core.js";
-import type { ErrorKind } from "./proxy-core.js";
-import type { RawHeaders } from "./types.js";
+import { getActiveProviders, insertRequestLog } from "../../db/index.js";
+import { getSetting } from "../../db/settings.js";
+import { decrypt } from "../../utils/crypto.js";
+import { proxyGetRequest, createErrorFormatter, type ProxyErrorResponse } from "../proxy-core.js";
+import type { ErrorKind } from "../proxy-core.js";
+import type { RawHeaders } from "../types.js";
 import { handleProxyRequest, type RouteHandlerDeps } from "./proxy-handler.js";
-import { createOrchestrator } from "./orchestration/orchestrator.js";
-import { RetryRuleMatcher } from "./orchestration/retry-rules.js";
-import { ProviderSemaphoreManager } from "./orchestration/semaphore.js";
-import type { RequestTracker } from "../monitor/request-tracker.js";
-import type { UsageWindowTracker } from "./routing/usage-window-tracker.js";
-import type { AdaptiveConcurrencyController } from "./adaptive-controller.js";
-import { HTTP_NOT_FOUND, HTTP_BAD_GATEWAY } from "../core/constants.js";
+import { createOrchestrator } from "../orchestration/orchestrator.js";
+import { RetryRuleMatcher } from "../orchestration/retry-rules.js";
+import { ProviderSemaphoreManager } from "../orchestration/semaphore.js";
+import type { RequestTracker } from "../../monitor/request-tracker.js";
+import type { UsageWindowTracker } from "../routing/usage-window-tracker.js";
+import type { AdaptiveConcurrencyController } from "../adaptive-controller.js";
+import { HTTP_NOT_FOUND, HTTP_BAD_GATEWAY } from "../../core/constants.js";
 
 export interface OpenaiProxyOptions {
   db: Database.Database;
@@ -25,7 +25,7 @@ export interface OpenaiProxyOptions {
   semaphoreManager?: ProviderSemaphoreManager;
   tracker?: RequestTracker;
   usageWindowTracker?: UsageWindowTracker;
-  sessionTracker?: import("./loop-prevention/session-tracker.js").SessionTracker;
+  sessionTracker?: import("../loop-prevention/session-tracker.js").SessionTracker;
   adaptiveController?: AdaptiveConcurrencyController;
 }
